+++ Common/src/main/java/net/merchantpug/apugli/access/ScaleDataAccess.java
@@ -1,0 +1,8 @@
+package net.merchantpug.apugli.access;
+
+import net.minecraft.resources.ResourceLocation;
+
+public interface ScaleDataAccess {
+    void apugli$addToApoliScaleModifiers(ResourceLocation value);
+    void apugli$removeFromApoliScaleModifiers(ResourceLocation value);
+}