+++ Fabric/src/main/java/net/merchantpug/apugli/ApugliFabric.java
@@ -1,6 +1,7 @@
 package net.merchantpug.apugli;
 
 import eu.midnightdust.lib.config.MidnightConfig;
+import io.github.apace100.apoli.Apoli;
 import io.github.apace100.apoli.integration.PostPowerLoadCallback;
 import io.github.apace100.apoli.integration.PrePowerReloadCallback;
 import io.github.apace100.apoli.power.Power;
@@ -8,8 +9,10 @@ import io.github.apace100.apoli.util.NamespaceAlias;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.loader.api.FabricLoader;
+import net.merchantpug.apugli.mixin.fabric.common.accessor.PowerTypeRegistryAccessor;
 import net.merchantpug.apugli.network.ApugliPackets;
 import net.merchantpug.apugli.network.s2c.UpdateUrlTexturesPacket;
+import net.merchantpug.apugli.network.s2c.integration.pehkui.ClearScaleModifierCachePacket;
 import net.merchantpug.apugli.power.CustomProjectilePower;
 import net.merchantpug.apugli.power.TextureOrUrlPower;
 import net.merchantpug.apugli.registry.power.ApugliPowers;
@@ -34,9 +37,23 @@ public class ApugliFabric implements ModInitializer {
         Apugli.init();
         ApugliPackets.registerC2S();
 
-        PrePowerReloadCallback.EVENT.register(() -> TextureUtil.getCache().clear());
+        ServerLifecycleEvents.START_DATA_PACK_RELOAD.register((server, serverResourceManager) -> {
+            if (FabricLoader.getInstance().isModLoaded("pehkui")) {
+                ApugliPowers.MODIFY_SCALE.get().clearFromAll();
+                ApugliPowers.MODIFY_SCALE.get().clearScaleTypeCache();
+                ApugliPowers.MODIFY_SCALE.get().clearModifiersFromCache();
+                server.getPlayerList().getPlayers().forEach(serverPlayer -> ApugliPackets.sendS2C(new ClearScaleModifierCachePacket(), serverPlayer));
+            }
+            TextureUtil.getCache().clear();
+        });
 
         PostPowerLoadCallback.EVENT.register((powerId, factoryId, isSubPower, json, powerType) -> {
+            if (!FabricLoader.getInstance().isModLoaded("pehkui") && factoryId.equals(Apugli.asResource("modify_scale"))) {
+                Apugli.LOG.error("Power '" + powerId + "' could not be loaded as it requires the Pehkui mod to be present. (skipping).");
+                PowerTypeRegistryAccessor.invokeRemove(powerId);
+                return;
+            }
+
             Power power = powerType.create(null);
             if ((power instanceof TextureOrUrlPower texturePower) && texturePower.getTextureUrl() != null) {
                 TextureUtil.cachePower(powerId, texturePower);